+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -310,6 +310,11 @@ New checks
 
   Warns if you use the ternary operator
 
+- New :doc:`bsl-type-decl-header
+  <clang-tidy/checks/bsl-type-decl-header>` check.
+
+  Checks that type name declared in header file identical to file name.
+
 - New :doc:`bsl-using-decl-scope
   <clang-tidy/checks/bsl-using-decl-scope>` check.